requirements.txt
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
openpyxl
pyyaml
dash
plotly
